+++ .gitignore
@@ -43,6 +43,7 @@ Thumbs.db
 /mapcache
 /nbproject
 /yaml2sql
+/yamlupgrade
 
 # /3rdparty/libconfig/
 /3rdparty/libconfig/Makefile
@@ -122,6 +123,7 @@ Thumbs.db
 /runserver.bat
 /serv.bat
 /yaml2sql.bat
+/yamlupgrade.bat
 
 # dlls
 /libmysql.dll
