+++ db/achievement_db.yml
@@ -16,70 +16,34 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 ###########################################################################
-# Renewal Achievement Database
+# Achievement Database
 ###########################################################################
 #
 # Achievement Settings
 #
 ###########################################################################
-# ID - Unique achievement ID.
-###########################################################################
-# Group - Achievement group type. Each achievement type calls a specific
-# objective check.
-# Valid groups:
-#  AG_ADD_FRIEND
-#  AG_ADVENTURE
-#  AG_BABY
-#  AG_BATTLE
-#  AG_CHATTING
-#  AG_CHATTING_COUNT
-#  AG_CHATTING_CREATE
-#  AG_CHATTING_DYING
-#  AG_EAT
-#  AG_GET_ITEM
-#  AG_GET_ZENY
-#  AG_GOAL_ACHIEVE
-#  AG_GOAL_LEVEL
-#  AG_GOAL_STATUS
-#  AG_HEAR
-#  AG_JOB_CHANGE
-#  AG_MARRY
-#  AG_PARTY
-#  AG_ENCHANT_FAIL
-#  AG_ENCHANT_SUCCESS
-#  AG_SEE
-#  AG_SPEND_ZENY
-#  AG_TAMING
-###########################################################################
-# Name - Achievement name. Used when sending rewards through RODEX.
-###########################################################################
-# Target - A list of monster ID and count values that the achievement
-# requires. The target count can also be used for achievements that keep
-# a counter while not being related to monster kills.
-# Capped at MAX_ACHIEVEMENT_OBJECTIVES.
-###########################################################################
-# Condition - A conditional statement that must be met for the achievement
-# to be considered complete.
-###########################################################################
-# Map - A map name that is used for the AG_CHATTING type which increments
-# the counter based on the player's map.
-###########################################################################
-# Dependent: - A list of achievement IDs that need to be completed before
-# this achievement is considered complete.
-###########################################################################
-# Reward - A list of rewards that are given on completion. All fields are
-# optional.
-#   ItemID: Item ID
-#   Amount:  Amount of Item ID (default 1)
-#   Script: Bonus Script
-#   TitleID: Title ID
-###########################################################################
-# Score - Achievement points that are given on completion.
+# - Id                  Achievement ID.
+#   Group               Achievement group type. (Defaut: None)
+#   Name                Achievement name.
+#   Targets:            List of targets the achievement requires. (Default: null)
+#     - Id              Index value used for import methods.
+#       Mob             Target mob. (Default: 0)
+#       Count           Target count. (Default: 1)
+#   Condition           Conditional statement that must be met for the achievement to be considered complete. (Default: null)
+#   Map                 Map name that is used for the AG_CHATTING type. (Default: -1)
+#   Dependents:         List of achievements that need to be completed before this achievement is considered complete. (Default: null)
+#     - Id: <bool>      Achievement ID pre-requisite.
+#   Rewards:            List of rewards that are given on completion. (Default: null)
+#     Item              Item name.
+#     Amount            Amount of item. (Default: 1)
+#     Script            Bonus Script. (Default: null)
+#     TitleId           Title ID. (Default: 0)
+#   Score               Achievement points that are given on completion. (Default: 0)
 ###########################################################################
 
 Header:
   Type: ACHIEVEMENT_DB
-  Version: 1
+  Version: 2
 
 Footer:
   Imports:
