+++ db/pre-re/achievement_level_db.yml
@@ -16,15 +16,14 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 ###########################################################################
-# Pre-Renewal Achievement Level Database
+# Achievement Level Database
 ###########################################################################
 #
 # Achievement Level Settings
 #
 ###########################################################################
-# Level - Achievement Level
-###########################################################################
-# Points - Required total scoring points to reach this level.
+# - Level     Achievement Level.
+#   Points    Required total scoring points to reach this level.
 ###########################################################################
 
 Header:
