+++ doc/achievements.txt
@@ -3,59 +3,60 @@
 //===== By: ==================================================
 //= rAthena Dev Team
 //===== Last Updated: ========================================
-//= 20190226
+//= 20200220
 //===== Description: =========================================
 //= Explanation of the achievements_db.yml file and structure.
 //============================================================
 
 ---------------------------------------
 
-ID: Unique achievement ID.
+Id: Unique achievement ID.
 
 ---------------------------------------
 
 Group: Achievement group type. Each achievement type calls a specific objective check.
 Valid groups:
-	AG_ADD_FRIEND - Triggered when a player adds a friend.
-	AG_ADVENTURE - Does not trigger automatically. These are triggered by the achievementcomplete script command.
-	AG_BABY - Triggered when a player becomes a baby job.
-	AG_BATTLE - Triggered when a player kills a monster.
-	AG_CHATTING - Unknown.
-	AG_CHATTING_COUNT - Triggered when a player has a chatroom open and others join.
-	AG_CHATTING_CREATE - Triggered when a player creates a chatroom.
-	AG_CHATTING_DYING - Triggered when a player creates a chatroom and dies with it open.
-	AG_EAT - Unknown.
-	AG_GET_ITEM - Triggered when a player gets an item that has a specific sell value.
-	AG_GET_ZENY - Triggered when a player gets a specific amount of zeny at once.
-	AG_GOAL_ACHIEVE - Triggered when a player's achievement rank levels up.
-	AG_GOAL_LEVEL - Triggered when a player's base level or job level changes.
-	AG_GOAL_STATUS - Triggered when a player's base stats changes.
-	AG_HEAR - Unknown.
-	AG_JOB_CHANGE - Triggered when a player's job changes.
-	AG_MARRY - Triggered when two players get married.
-	AG_PARTY - Triggered when a player creates a party.
-	AG_ENCHANT_FAIL - Triggered when a player fails to refine an equipment.
-	AG_ENCHANT_SUCCESS - Triggered when a player successfully refines an equipment.
-	AG_SEE - Unknown.
-	AG_SPEND_ZENY - Triggered when a player spends any amount of zeny on vendors.
-	AG_TAMING - Triggered when a player tames a monster.
+	None - Can be used for custom achievements that are given through a script with no trigger events.
+	Add_Friend - Triggered when a player adds a friend.
+	Adventure - Does not trigger automatically. These are triggered by the achievementcomplete script command.
+	Baby - Triggered when a player becomes a baby job.
+	Battle - Triggered when a player kills a monster.
+	Chatting - Aegis uses this when talking to a NPC. These are triggered by the achievementupdate script command.
+	Chatting_Count - Triggered when a player has a chatroom open and others join.
+	Chatting_Create - Triggered when a player creates a chatroom.
+	Chatting_Dying - Triggered when a player creates a chatroom and dies with it open.
+	Eat - Unknown.
+	Get_Item - Triggered when a player gets an item that has a specific sell value.
+	Get_Zeny - Triggered when a player gets a specific amount of zeny at once.
+	Goal_Achieve - Triggered when a player's achievement rank levels up.
+	Goal_Level - Triggered when a player's base level or job level changes.
+	Goal_Status - Triggered when a player's base stats changes.
+	Job_Change - Triggered when a player's job changes.
+	Marry - Triggered when two players get married.
+	Party - Triggered when a player creates a party.
+	Enchant_Fail - Triggered when a player fails to refine an equipment.
+	Enchant_Success - Triggered when a player successfully refines an equipment.
+	Spend_Zeny - Triggered when a player spends any amount of zeny on vendors.
+	Taming - Triggered when a player tames a monster.
 
 ---------------------------------------
 
-Name: Achievement name. Not read into source but used for quick look ups.
+Name: Achievement name. Used when sending rewards through RODEX.
 
 ---------------------------------------
 
-Target: A list of monster ID and count values that the achievement requires.
+Targets: A list of monster names and count values that the achievement requires.
 		The target count can also be used for achievements that keep a counter while not being related to monster kills.
 		Capped at MAX_ACHIEVEMENT_OBJECTIVES.
 
 Example:
   // Player must kill 5 Scorpions and 10 Poring.
-  Target:
-    - MobID: 1001
+  Targets:
+    - Id: 0
+      Mob: SCORPION
       Count: 5
-    - MobID: 1002
+    - Id: 1
+      Mob: PORING
       Count: 10
 
 Example 2:
@@ -63,7 +64,7 @@ Example 2:
   // and not checked for a total (UI_Type = 1).
   // IE: In the achievement_list.lub file, UI_Type 0 is displayed as non-incremental while 1 shows a progress bar of completion for the achievement.
   Condition: " ARG0 >= 100 "
-  Target:
+  Targets:
     - Id: 0 // Array index value
       Count: 100
 
@@ -84,19 +85,30 @@ Example 2:
 
 ---------------------------------------
 
-Map: A map name that is used for the AG_CHATTING type which increments the counter based on the player's map.
+Map: A map name that is used for the Chatting group which increments the counter based on the player's map.
+NOTICE: This option is currently disabled until the official behavior is confirmed.
 
 ---------------------------------------
 
-Dependent: A list of achievement IDs that need to be completed before this achievement is considered complete.
+Dependents: A list of achievement IDs that need to be completed before this achievement is considered complete.
+
+Example:
+  // Player must complete achievements 10001 and 10002 first.
+  Dependents:
+    10001: true
+    10002: true
+
+  // Used with the import, dependent achievements can be disabled. The player now only requires completion of achievement 10001.
+  Dependents:
+    10002: false
 
 ---------------------------------------
 
-Reward: A list of rewards that are given on completion. All fields are optional.
-  ItemID: Item ID
-  Amount:	Amount of Item ID (default 1)
+Rewards: A list of rewards that are given on completion. All fields are optional.
+  Item: Item Name
+  Amount: Amount of Item (Default: 1)
   Script: Bonus Script
-  TitleID: Title ID
+  TitleId: Title ID
 
 ---------------------------------------
