+++ src/map/achievement.cpp
@@ -50,8 +50,7 @@ const std::string AchievementDatabase::getDefaultLocation(){
 uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 	uint32 achievement_id;
 
-	// TODO: doesnt match camel case
-	if( !this->asUInt32( node, "ID", achievement_id ) ){
+	if( !this->asUInt32( node, "Id", achievement_id ) ){
 		return 0;
 	}
 
@@ -59,11 +58,7 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 	bool exists = achievement != nullptr;
 
 	if( !exists ){
-		if( !this->nodeExists( node, "Group" ) ){
-			return 0;
-		}
-
-		if( !this->nodeExists( node, "Name" ) ){
+		if( !this->nodesExist( node, { "Name" } ) ){
 			return 0;
 		}
 
@@ -78,14 +73,18 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 			return 0;
 		}
 
+		std::string group_name_constant = "AG_" + group_name;
 		int64 constant;
 
-		if( !script_get_constant( group_name.c_str(), &constant ) ){
-			this->invalidWarning( node, "achievement_read_db_sub: Invalid group %s for achievement %d, skipping.\n", group_name.c_str(), achievement_id );
+		if( !script_get_constant( group_name_constant.c_str(), &constant ) ){
+			this->invalidWarning( node, "Invalid Group %s for achievement %d, skipping.\n", group_name.c_str(), achievement_id );
 			return 0;
 		}
 
 		achievement->group = (e_achievement_group)constant;
+	} else {
+		if (!exists)
+			achievement->group = AG_NONE;
 	}
 
 	if( this->nodeExists( node, "Name" ) ){
@@ -98,15 +97,10 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 		achievement->name = name;
 	}
 
-	if( this->nodeExists( node, "Target" ) ){
-		const YAML::Node& targets = node["Target"];
+	if( this->nodeExists( node, "Targets" ) ){
+		const YAML::Node& targets = node["Targets"];
 
 		for( const YAML::Node& targetNode : targets ){
-			if( achievement->targets.size() >= MAX_ACHIEVEMENT_OBJECTIVES ){
-				this->invalidWarning( targetNode, "Node \"Target\" list exceeds the maximum of %d, skipping.\n", MAX_ACHIEVEMENT_OBJECTIVES );
-				return 0;
-			}
-
 			uint16 targetId;
 
 			if( !this->asUInt16( targetNode, "Id", targetId ) ){
@@ -114,16 +108,16 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 			}
 
 			if( targetId >= MAX_ACHIEVEMENT_OBJECTIVES ){
-				this->invalidWarning( targetNode["Id"], "Node \"Id\" is out of valid range [0,%d], skipping.\n", MAX_ACHIEVEMENT_OBJECTIVES );
-				return 0;
+				this->invalidWarning( targetNode["Id"], "Target Id is out of valid range [0,%d], skipping.\n", MAX_ACHIEVEMENT_OBJECTIVES );
+				continue;
 			}
 
 			std::shared_ptr<achievement_target> target = rathena::util::map_find( achievement->targets, targetId );
 			bool targetExists = target != nullptr;
 
 			if( !targetExists ){
-				if( !this->nodeExists( targetNode, "Count" ) && !this->nodeExists( targetNode, "MobID" ) ){
-					this->invalidWarning( targetNode, "Node \"Target\" has no data specified, skipping.\n" );
+				if( !this->nodeExists( targetNode, "Count" ) && !this->nodeExists( targetNode, "Mob" ) ){
+					this->invalidWarning( targetNode, "Target has no data specified, skipping.\n" );
 					return 0;
 				}
 
@@ -137,30 +131,43 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 					return 0;
 				}
 
+				if( count == 0 ){
+					if( targetExists ){
+						achievement->targets.erase( targetId );
+						continue;
+					}else{
+						this->invalidWarning( targetNode["Count"], "Target count has to be > 0, skipping.\n" );
+						return 0;
+					}
+				}
+
 				target->count = count;
 			}else{
 				if( !targetExists ){
-					target->count = 0;
+					target->count = 1;
 				}
 			}
 
-			if( this->nodeExists( targetNode, "MobID" ) ){
+			if( this->nodeExists( targetNode, "Mob" ) ){
 				if( achievement->group != AG_BATTLE && achievement->group != AG_TAMING ){
-					this->invalidWarning( targets, "Node \"MobID\" is only supported for targets in group AG_BATTLE or AG_TAMING, skipping.\n" );
+					this->invalidWarning( targets, "Target Mob is only supported for targets in group AG_BATTLE or AG_TAMING, skipping.\n" );
+					continue;
+				}
+
+				std::string mob_name;
+
+				if( !this->asString( targetNode, "Mob", mob_name ) ){
 					return 0;
 				}
 
-				uint32 mob_id;
+				std::shared_ptr<s_mob_db> mob = mobdb_search_aegisname( mob_name.c_str() );
 
-				// TODO: not camel case
-				if( !this->asUInt32( targetNode, "MobID", mob_id ) ){
+				if (mob == nullptr) {
+					this->invalidWarning(targetNode["Mob"], "Target Mob %s does not exist, skipping.\n", mob_name.c_str());
 					return 0;
 				}
 
-				if( mob_db.find( mob_id ) == nullptr ){
-					this->invalidWarning( targetNode["MobID"], "Unknown monster ID %d, skipping.\n", mob_id );
-					return 0;
-				}
+				uint32 mob_id = mob->vd.class_;
 
 				if( !this->mobexists( mob_id ) ){
 					this->achievement_mobs.push_back( mob_id );
@@ -195,12 +202,13 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 
 		achievement->condition = parse_script( condition.c_str(), this->getCurrentFile().c_str(), node["Condition"].Mark().line + 1, SCRIPT_IGNORE_EXTERNAL_BRACKETS );
 	}else{
-		achievement->condition = nullptr;
+		if (!exists)
+			achievement->condition = nullptr;
 	}
 
 	if( this->nodeExists( node, "Map" ) ){
-		if( achievement->group != AG_CHAT ){
-			this->invalidWarning( node, "Node \"Map\" can only be used with the group AG_CHATTING, skipping.\n" );
+		if( achievement->group != AG_CHATTING ){
+			this->invalidWarning( node, "Map can only be used with the group AG_CHATTING, skipping.\n" );
 			return 0;
 		}
 
@@ -213,7 +221,7 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 		achievement->mapindex = map_mapname2mapid( mapname.c_str() );
 
 		if( achievement->mapindex == -1 ){
-			this->invalidWarning( node["Map"], "Unknown map name '%s'.\n", mapname.c_str() );
+			this->invalidWarning( node["Map"], "Map %s does not exist, skipping.\n", mapname.c_str() );
 			return 0;
 		}
 	}else{
@@ -222,43 +230,51 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 		}
 	}
 
-	if( this->nodeExists( node, "Dependent" ) ){
-		for( const YAML::Node& subNode : node["Dependent"] ){
-			uint32 dependent_achievement_id;
+	if( this->nodeExists( node, "Dependents" ) ){
+		const YAML::Node &dependentNode = node["Dependents"];
 
-			if( !this->asUInt32( subNode, "Id", dependent_achievement_id ) ){
+		for( const auto it : dependentNode ){
+			uint32 dependent_achievement_id = it.first.as<uint32>();
+			bool active;
+
+			if (!this->asBool(dependentNode, std::to_string(dependent_achievement_id), active))
 				return 0;
-			}
 
-			// TODO: import logic for clearing => continue
-			// TODO: change to set to prevent multiple entries with the same id?
-			achievement->dependent_ids.push_back( dependent_achievement_id );
+			if (active) {
+				if (std::find(achievement->dependent_ids.begin(), achievement->dependent_ids.end(), dependent_achievement_id) != achievement->dependent_ids.end()) {
+					this->invalidWarning(dependentNode, "Dependent achievement %d is already part of the list, skipping.\n", dependent_achievement_id);
+					continue;
+				}
+
+				if (achievement->dependent_ids.size() >= MAX_ACHIEVEMENT_DEPENDENTS) {
+					this->invalidWarning(dependentNode, "Maximum amount (%d) of dependent achievements reached, skipping.\n", MAX_ACHIEVEMENT_DEPENDENTS);
+					break;
+				}
+
+				achievement->dependent_ids.push_back(dependent_achievement_id);
+			} else
+				util::vector_erase_if_exists(achievement->dependent_ids, dependent_achievement_id);
 		}
 	}
 
-	// TODO: not plural
-	if( this->nodeExists( node, "Reward" ) ){
-		const YAML::Node& rewardNode = node["Reward"];
+	if( this->nodeExists( node, "Rewards" ) ){
+		const YAML::Node& rewardNode = node["Rewards"];
 
-		// TODO: not camel case
-		if( this->nodeExists( rewardNode, "ItemID" ) ){
-			t_itemid itemId;
+		if( this->nodeExists( rewardNode, "Item" ) ){
+			std::string item_name;
 
-			if( !this->asUInt32( rewardNode, "ItemID", itemId ) ){
+			if( !this->asString( rewardNode, "Item", item_name ) ){
 				return 0;
 			}
 
-			if( !itemdb_exists( itemId ) ){
-				this->invalidWarning( rewardNode["ItemID"], "Unknown item with ID %u.\n", itemId );
+			struct item_data *item = itemdb_search_aegisname(item_name.c_str());
+
+			if (item == nullptr) {
+				this->invalidWarning(rewardNode["Item"], "Reward Item %s does not exist, skipping.\n", item_name.c_str());
 				return 0;
 			}
 
-			achievement->rewards.nameid = itemId;
-
-			if( achievement->rewards.amount == 0 ){
-				// Default the amount to 1
-				achievement->rewards.amount = 1;
-			}
+			achievement->rewards.nameid = item->nameid;
 		}
 
 		if( this->nodeExists( rewardNode, "Amount" ) ){
@@ -269,6 +285,9 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 			}
 
 			achievement->rewards.amount = amount;
+		} else {
+			if (!exists)
+				achievement->rewards.amount = 1;
 		}
 
 		if( this->nodeExists( rewardNode, "Script" ) ){
@@ -285,18 +304,26 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 
 			achievement->rewards.script = parse_script( script.c_str(), this->getCurrentFile().c_str(), achievement_id, SCRIPT_IGNORE_EXTERNAL_BRACKETS );
 		}else{
-			achievement->rewards.script = nullptr;
+			if (!exists)
+				achievement->rewards.script = nullptr;
 		}
 
-		// TODO: not camel case
-		if( this->nodeExists( rewardNode, "TitleID" ) ){
+		if( this->nodeExists( rewardNode, "TitleId" ) ){
 			uint32 title;
 
-			if( !this->asUInt32( rewardNode, "TitleID", title ) ){
+			if( !this->asUInt32( rewardNode, "TitleId", title ) ){
+				return 0;
+			}
+
+			if (title < TITLE_BASE || title > TITLE_MAX) {
+				this->invalidWarning(rewardNode["TitleId"], "Reward Title ID %u does not exist (%hu~%hu), skipping.\n", title, TITLE_BASE, TITLE_MAX);
 				return 0;
 			}
 
 			achievement->rewards.title_id = title;
+		} else {
+			if (!exists)
+				achievement->rewards.title_id = 0;
 		}
 	}
 
@@ -308,6 +335,9 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 		}
 
 		achievement->score = score;
+	} else {
+		if (!exists)
+			achievement->score = 0;
 	}
 
 	if( !exists ){
@@ -315,6 +345,25 @@ uint64 AchievementDatabase::parseBodyNode(const YAML::Node &node){
 	}
 
 	return 1;
+}
+
+void AchievementDatabase::loadingFinished(){
+	for (const auto &achit : achievement_db) {
+		const std::shared_ptr<s_achievement_db> ach = achit.second;
+
+		for (auto dep = ach->dependent_ids.begin(); dep != ach->dependent_ids.end(); dep++) {
+			if (!achievement_db.exists(*dep)) {
+				ShowWarning("achievement_read_db: An invalid Dependent ID %d was given for Achievement %d. Removing from list.\n", *dep, ach->achievement_id);
+				dep = ach->dependent_ids.erase(dep);
+
+				if (dep == ach->dependent_ids.end()) {
+					break;
+				}
+			}
+		}
+
+		ach->dependent_ids.shrink_to_fit();
+	}
 }
 
 AchievementDatabase achievement_db;
@@ -507,8 +556,8 @@ bool achievement_check_dependent(struct map_session_data *sd, int achievement_id
 
 	// Check if the achievement has a dependent
 	// If so, then do a check on all dependents to see if they're complete
-	for (int i = 0; i < adb->dependent_ids.size(); i++) {
-		if (!achievement_done(sd, adb->dependent_ids[i]))
+	for (const auto &depit : adb->dependent_ids) {
+		if (!achievement_done(sd, depit))
 			return false; // One of the dependent is not complete!
 	}
 
@@ -851,6 +900,19 @@ bool achievement_check_condition( struct script_code* condition, struct map_sess
 }
 
 /**
+ * Check to see if an achievement's target count is complete
+ * @param ad: Achievement data
+ * @param current_count: Current target data
+ * @return True if all target values meet the requirements or false otherwise
+ */
+static bool achievement_target_complete(std::shared_ptr<s_achievement_db> ad, std::array<int, MAX_ACHIEVEMENT_OBJECTIVES> current_count) {
+	return std::find_if(ad->targets.begin(), ad->targets.end(),
+		[current_count](const std::pair<uint16, std::shared_ptr<achievement_target>> &target) -> bool {
+		return current_count[target.first] < target.second->count;
+	}) == ad->targets.end();
+}
+
+/**
  * Update achievement objectives.
  * @param sd: Player to update
  * @param ad: Achievement data to compare for completion
@@ -892,9 +954,9 @@ static bool achievement_update_objectives(struct map_session_data *sd, std::shar
 	switch (group) {
 		case AG_ADD_FRIEND:
 		case AG_BABY:
-		case AG_CHAT_COUNT:
-		case AG_CHAT_CREATE:
-		case AG_CHAT_DYING:
+		case AG_CHATTING_COUNT:
+		case AG_CHATTING_CREATE:
+		case AG_CHATTING_DYING:
 		case AG_GET_ITEM:
 		case AG_GET_ZENY:
 		case AG_GOAL_LEVEL:
@@ -902,8 +964,8 @@ static bool achievement_update_objectives(struct map_session_data *sd, std::shar
 		case AG_JOB_CHANGE:
 		case AG_MARRY:
 		case AG_PARTY:
-		case AG_REFINE_FAIL:
-		case AG_REFINE_SUCCESS:
+		case AG_ENCHANT_FAIL:
+		case AG_ENCHANT_SUCCESS:
 			if (!ad->condition)
 				return false;
 
@@ -914,14 +976,8 @@ static bool achievement_update_objectives(struct map_session_data *sd, std::shar
 			complete = true;
 			break;
 		case AG_SPEND_ZENY:
-		//case AG_CHAT: // No information on trigger events
 			if (ad->targets.empty() || !ad->condition)
 				return false;
-
-			//if (group == AG_CHAT) {
-			//	if (ad->mapindex > -1 && sd->bl.m != ad->mapindex)
-			//		return false;
-			//}
 
 			for (const auto &it : ad->targets) {
 				if (current_count[it.first] < it.second->count)
@@ -933,11 +989,7 @@ static bool achievement_update_objectives(struct map_session_data *sd, std::shar
 
 			changed = true;
 
-			if (std::find_if(ad->targets.begin(), ad->targets.end(),
-				[current_count](const std::pair<uint16, std::shared_ptr<achievement_target>> &target) -> bool {
-					return current_count[target.first] < target.second->count;
-				}
-			) == ad->targets.end())
+			if (achievement_target_complete(ad, current_count))
 				complete = true;
 			break;
 		case AG_BATTLE:
@@ -955,13 +1007,38 @@ static bool achievement_update_objectives(struct map_session_data *sd, std::shar
 			if (!changed)
 				return false;
 
-			if (std::find_if(ad->targets.begin(), ad->targets.end(),
-				[current_count](const std::pair<uint16, std::shared_ptr<achievement_target>> &target) -> bool {
-					return current_count[target.first] < target.second->count;
-				}
-			) == ad->targets.end())
+			if (achievement_target_complete(ad, current_count))
 				complete = true;
 			break;
+		case AG_GOAL_ACHIEVE:
+			if (!achievement_check_condition(ad->condition, sd)) // Parameters weren't met
+				return false;
+
+			changed = true;
+			complete = true;
+			break;
+		/*
+		case AG_CHATTING:
+			if (ad->targets.empty())
+				return false;
+
+			if (ad->mapindex > -1 && sd->bl.m != ad->mapindex)
+				return false;
+
+			for (const auto &it : ad->targets) {
+				if (current_count[it.first] < it.second->count) {
+					current_count[it.first]++;
+					changed = true;
+				}
+			}
+
+			if (!changed)
+				return false;
+
+			if (achievement_target_complete(ad, current_count))
+				complete = true;
+			break;
+		*/
 	}
 
 	if( isNew ){
@@ -1022,15 +1099,8 @@ void achievement_update_objective(struct map_session_data *sd, enum e_achievemen
 		}
 		va_end(ap);
 
-		switch(group) {
-			case AG_CHAT: //! TODO: Not sure how this works officially
-				// These have no objective use.
-				break;
-			default:
-				for (auto &ach : achievement_db)
-					achievement_update_objectives(sd, ach.second, group, count);
-				break;
-		}
+		for (auto &ach : achievement_db)
+			achievement_update_objectives(sd, ach.second, group, count);
 
 		// Remove variables that might have been set
 		for (int i = 0; i < arg_count; i++){
@@ -1070,33 +1140,10 @@ int achievement_update_objective_sub(block_list *bl, va_list ap)
 }
 
 /**
- * Loads achievements from the achievement db.
- */
-void achievement_read_db(void)
-{	
-	achievement_db.load();
-
-	for (auto &achit : achievement_db) {
-		const auto ach = achit.second;
-
-		for (int i = 0; i < ach->dependent_ids.size(); i++) {
-			if (!achievement_db.exists(ach->dependent_ids[i])) {
-				ShowWarning("achievement_read_db: An invalid Dependent ID %d was given for Achievement %d. Removing from list.\n", ach->dependent_ids[i], ach->achievement_id);
-				ach->dependent_ids.erase(ach->dependent_ids.begin() + i);
-			}
-		}
-	}
-
-	achievement_level_db.load();
-}
-
-/**
  * Reloads the achievement database
  */
 void achievement_db_reload(void)
 {
-	if (!battle_config.feature_achievement)
-		return;
 	do_final_achievement();
 	do_init_achievement();
 }
@@ -1108,7 +1155,8 @@ void do_init_achievement(void)
 {
 	if (!battle_config.feature_achievement)
 		return;
-	achievement_read_db();
+	achievement_db.load();
+	achievement_level_db.load();
 }
 
 /**
@@ -1129,7 +1177,7 @@ s_achievement_db::s_achievement_db()
 	, targets()
 	, dependent_ids()
 	, condition(nullptr)
-	, mapindex(0)
+	, mapindex(-1)
 	, rewards()
 	, score(0)
 	, has_dependent(0)
