+++ src/tool/yamlupgrade.hpp
@@ -1,0 +1,9 @@
+// Copyright (c) rAthena Dev Teams - Licensed under GNU GPL
+// For more information, see LICENCE in the main folder
+
+#ifndef YAMLUPGRADE_HPP
+#define YAMLUPGRADE_HPP
+
+#include "yaml.hpp"
+
+#endif /* YAMLUPGRADE_HPP */
