+++ src/tool/yamlupgrade.vcxproj
@@ -19,9 +19,9 @@
     </ProjectConfiguration>
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{5A9059F2-4933-49A2-BEE6-CC67F66FA070}</ProjectGuid>
+    <ProjectGuid>{9115C6D1-520A-4540-B4FE-95F3C923FE2C}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
-    <RootNamespace>csv2yaml</RootNamespace>
+    <RootNamespace>yamlupgrade</RootNamespace>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -159,12 +159,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="csv2yaml.hpp" />
     <ClInclude Include="yaml.hpp" />
+    <ClInclude Include="yamlupgrade.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\common\database.cpp" />
-    <ClCompile Include="csv2yaml.cpp" />
+    <ClCompile Include="yamlupgrade.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\common\database.hpp" />
@@ -175,11 +175,11 @@
   <Target Name="AfterClean">
     <Delete Files="$(SolutionDir)zlib.dll" ContinueOnError="true" />
     <Delete Files="$(SolutionDir)serv.bat" ContinueOnError="true" />
-    <Delete Files="$(SolutionDir)csv2yaml.bat" ContinueOnError="true" />
+    <Delete Files="$(SolutionDir)yamlupgrade.bat" ContinueOnError="true" />
   </Target>
   <Target Name="AfterBuild">
     <Copy SourceFiles="$(SolutionDir)3rdparty\zlib\lib\$(Platform)\zlib.dll" DestinationFolder="$(SolutionDir)" ContinueOnError="true" Condition="!Exists('$(SolutionDir)zlib.dll')" />
     <Copy SourceFiles="$(SolutionDir)tools\serv.bat" DestinationFolder="$(SolutionDir)" ContinueOnError="true" Condition="!Exists('$(SolutionDir)serv.bat')" />
-    <Copy SourceFiles="$(SolutionDir)tools\csv2yaml.bat" DestinationFolder="$(SolutionDir)" ContinueOnError="true" Condition="!Exists('$(SolutionDir)csv2yaml.bat')" />
+    <Copy SourceFiles="$(SolutionDir)tools\yamlupgrade.bat" DestinationFolder="$(SolutionDir)" ContinueOnError="true" Condition="!Exists('$(SolutionDir)yamlupgrade.bat')" />
   </Target>
 </Project>
