+++ tools/yamlupgrade.bat
@@ -1,0 +1,2 @@
+@ECHO OFF
+CALL serv.bat yamlupgrade.exe YAML UPGRADE
